+++ packages/input/src/Component.tsx
@@ -14,7 +14,7 @@ import cn from 'classnames';
 
 import { Button } from '@alfalab/core-components-button';
 import { FormControl } from '@alfalab/core-components-form-control';
-import { Badge } from '@alfalab/core-components/badge';
+import { Badge } from '@alfalab/core-components-badge';
 import { useFocus } from '@alfalab/hooks';
 import { CrossCircleMIcon } from '@alfalab/icons-glyph/CrossCircleMIcon';
 import { CheckmarkCircleMIcon } from '@alfalab/icons-glyph/CheckmarkCircleMIcon';