+++ packages/tooltip/src/index.ts
@@ -1,1 +1,3 @@
-export * from './Component';
+export { TooltipDesktop as Tooltip } from './desktop';
+export type { TooltipDesktopProps as TooltipProps } from './desktop';
+export * from './component.responsive';