+++ packages/notification-manager/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 5.1.9
+
+### Patch Changes
+
+### [#418](https://github.com/core-ds/core-components/pull/418)
+
+-   Исправлена проблема с default-импортом в cjs форматах
+
+-   Обновлены зависимости
+    -   notification@6.1.9
+    -   portal@3.1.2
+    -   stack@4.0.2
+
 ## 5.1.8
 
 ### Patch Changes